+++ core/liquidity/supplied/statevar.go
@@ -24,9 +24,9 @@ import (
 type probabilityOfTradingConverter struct{}
 
 type probabilityOfTrading struct {
-	bidOffset      []num.Decimal // represents offsets from best bid
+	bidOffset      []uint32      // represents offsets from best bid - not sent over the wire
 	bidProbability []num.Decimal // probability[0] = probability of the best bid, probabtility[len-1] = probability of the worst bid
-	askOffset      []num.Decimal // represents offsets from best ask
+	askOffset      []uint32      // represents offsets from best ask - not sent over the wire
 	askProbability []num.Decimal // probability[0] = probability of the best ask, probabtility[len-1] = probability of the worst ask
 }
 
@@ -34,20 +34,20 @@ var (
 	defaultInRangeProbabilityOfTrading = num.DecimalFromFloat(0.5)
 	defaultMinimumProbabilityOfTrading = num.DecimalFromFloat(1e-8)
 	tolerance                          = num.DecimalFromFloat(1e-6)
-	incrementInPct                     = num.DecimalFromFloat(0.1)                                    // we calculate the probability of trading in increments of 0.1% of the best bid/ask
-	IncrementFactor                    = incrementInPct.Div(num.DecimalFromFloat(100))                // we calculate the probability of trading in increments of 0.001 from of the best bid/ask
-	maxDistanceWhenNoConsensusPct      = num.DecimalFromFloat(20)                                     // if there's no consensus yet and the price is within 20% of the best bid/ask it gets the default probability
-	maxDistanceWhenNoConsensusFactor   = maxDistanceWhenNoConsensusPct.Div(num.DecimalFromFloat(100)) // if there's no consensus yet and the price is within 0.2 of the best bid/ask it gets the default probability
+	OffsetIncrement                    = uint32(1000) // the increment of offsets in the offset slice - 1/PriceIncrementFactor as integer
+	OffsetIncrementAsDecimal           = num.DecimalFromInt64(int64(OffsetIncrement))
+	OffsetOneDecimal                   = OffsetIncrementAsDecimal.Mul(OffsetIncrementAsDecimal)             // an offset of 100%
+	PriceIncrementFactor               = num.DecimalOne().Div(num.DecimalFromInt64(int64(OffsetIncrement))) // we calculate the probability of trading in increments of 0.001 from of the best bid/ask
+	maxDistanceWhenNoConsensusPct      = num.DecimalFromFloat(20)                                           // if there's no consensus yet and the price is within 20% of the best bid/ask it gets the default probability
+	maxDistanceWhenNoConsensusFactor   = maxDistanceWhenNoConsensusPct.Div(num.DecimalFromFloat(100))       // if there's no consensus yet and the price is within 0.2 of the best bid/ask it gets the default probability
 
 	maxAskMultiplier = num.DecimalFromFloat(6) // arbitrarily large factor on the best ask - 600%
 )
 
 func (probabilityOfTradingConverter) BundleToInterface(kvb *statevar.KeyValueBundle) statevar.StateVariableResult {
 	return &probabilityOfTrading{
-		bidOffset:      kvb.KVT[0].Val.(*statevar.DecimalVector).Val,
-		bidProbability: kvb.KVT[1].Val.(*statevar.DecimalVector).Val,
-		askOffset:      kvb.KVT[2].Val.(*statevar.DecimalVector).Val,
-		askProbability: kvb.KVT[3].Val.(*statevar.DecimalVector).Val,
+		bidProbability: kvb.KVT[0].Val.(*statevar.DecimalVector).Val,
+		askProbability: kvb.KVT[1].Val.(*statevar.DecimalVector).Val,
 	}
 }
 
@@ -55,9 +55,7 @@ func (probabilityOfTradingConverter) InterfaceToBundle(res statevar.StateVariabl
 	value := res.(*probabilityOfTrading)
 	return &statevar.KeyValueBundle{
 		KVT: []statevar.KeyValueTol{
-			{Key: "bidOffset", Val: &statevar.DecimalVector{Val: value.bidOffset}, Tolerance: tolerance},
 			{Key: "bidProbability", Val: &statevar.DecimalVector{Val: value.bidProbability}, Tolerance: tolerance},
-			{Key: "askOffset", Val: &statevar.DecimalVector{Val: value.askOffset}, Tolerance: tolerance},
 			{Key: "askProbability", Val: &statevar.DecimalVector{Val: value.askProbability}, Tolerance: tolerance},
 		},
 	}
@@ -80,13 +78,11 @@ func (e *Engine) startCalcProbOfTrading(eventID string, endOfCalcCallback statev
 	}
 
 	// calculate offsets and probabilities for the range
-	bidOffsets, bidProbabilities := calculateBidRange(bestBid, IncrementFactor, tauScaled, e.rm.ProbabilityOfTrading)
-	askOffsets, askProbabilities := calculateAskRange(bestAsk, IncrementFactor, tauScaled, e.rm.ProbabilityOfTrading)
+	bidProbabilities := calculateBidRange(bestBid, PriceIncrementFactor, tauScaled, e.rm.ProbabilityOfTrading)
+	askProbabilities := calculateAskRange(bestAsk, PriceIncrementFactor, tauScaled, e.rm.ProbabilityOfTrading)
 
 	res := &probabilityOfTrading{
-		bidOffset:      bidOffsets,
 		bidProbability: bidProbabilities,
-		askOffset:      askOffsets,
 		askProbability: askProbabilities,
 	}
 	endOfCalcCallback.CalculationFinished(eventID, res, nil)
@@ -98,24 +94,20 @@ func (e *Engine) startCalcProbOfTrading(eventID string, endOfCalcCallback statev
 // and similarly probabilities[0] corresponds to best bid trading probability
 // whereas the last entry in offset equals to the maximum distance from bid for which
 // probability is calculated and is greater than the default minimum acceptable probability of trading.
-func calculateBidRange(bestBid, incrementFactor, tauScaled num.Decimal, probabilityFunc func(num.Decimal, num.Decimal, num.Decimal, num.Decimal, num.Decimal, bool, bool) num.Decimal) ([]num.Decimal, []num.Decimal) {
-	offsets := []num.Decimal{}
+func calculateBidRange(bestBid num.Decimal, priceIncrementFactor, tauScaled num.Decimal, probabilityFunc func(num.Decimal, num.Decimal, num.Decimal, num.Decimal, num.Decimal, bool, bool) num.Decimal) []num.Decimal {
 	probabilities := []num.Decimal{}
 
 	p := bestBid
-	offset := num.DecimalZero()
-	increment := incrementFactor.Mul(bestBid)
+	increment := priceIncrementFactor.Mul(bestBid)
 	for {
 		prob := probabilityFunc(bestBid, p, num.DecimalZero(), bestBid, tauScaled, true, true)
 		if prob.LessThanOrEqual(defaultMinimumProbabilityOfTrading) {
 			break
 		}
-		offsets = append(offsets, offset)
 		probabilities = append(probabilities, prob)
-		offset = offset.Add(incrementFactor)
 		p = p.Sub(increment)
 	}
-	return offsets, probabilities
+	return probabilities
 }
 
 // calculateAskRange calculates the probabilities of price between bestAsk and the worst ask (maxAsk)
@@ -125,30 +117,38 @@ func calculateBidRange(bestBid, incrementFactor, tauScaled num.Decimal, probabil
 // whereas the last entry in offset equals to the maximum distance from ask for which
 // probability is calculated, and the last entry in probabilities corresponds to probability of trading
 // at the price implied by this offset from best ask.
-func calculateAskRange(bestAsk, incrementFactor, tauScaled num.Decimal, probabilityFunc func(num.Decimal, num.Decimal, num.Decimal, num.Decimal, num.Decimal, bool, bool) num.Decimal) ([]num.Decimal, []num.Decimal) {
-	offsets := []num.Decimal{}
+func calculateAskRange(bestAsk num.Decimal, priceIncrementFactor, tauScaled num.Decimal, probabilityFunc func(num.Decimal, num.Decimal, num.Decimal, num.Decimal, num.Decimal, bool, bool) num.Decimal) []num.Decimal {
 	probabilities := []num.Decimal{}
 
 	maxAsk := bestAsk.Mul(maxAskMultiplier)
-	increment := incrementFactor.Mul(bestAsk)
+	increment := priceIncrementFactor.Mul(bestAsk)
+
 	p := bestAsk
-	offset := num.DecimalZero()
 	for {
 		prob := probabilityFunc(bestAsk, p, bestAsk, maxAsk, tauScaled, false, true)
 		if prob.LessThanOrEqual(defaultMinimumProbabilityOfTrading) {
 			break
 		}
-		offsets = append(offsets, offset)
 		probabilities = append(probabilities, prob)
-		offset = offset.Add(incrementFactor)
 		p = p.Add(increment)
 	}
-	return offsets, probabilities
+	return probabilities
 }
 
 // updatePriceBounds is called back from the state variable consensus engine when consensus is reached for the down/up factors and updates the price bounds.
 func (e *Engine) updateProbabilities(ctx context.Context, res statevar.StateVariableResult) error {
 	e.pot = res.(*probabilityOfTrading)
+
+	e.pot.bidOffset = make([]uint32, 0, len(e.pot.bidProbability))
+	for i := range e.pot.bidProbability {
+		e.pot.bidOffset = append(e.pot.bidOffset, uint32(i)*OffsetIncrement)
+	}
+
+	e.pot.askOffset = make([]uint32, 0, len(e.pot.bidProbability))
+	for i := range e.pot.askProbability {
+		e.pot.askOffset = append(e.pot.askOffset, uint32(i)*OffsetIncrement)
+	}
+
 	if e.log.GetLevel() <= logging.DebugLevel {
 		e.log.Debug("consensus reached for probability of trading", logging.String("market", e.marketID))
 	}
@@ -164,7 +164,7 @@ func (e *Engine) updateProbabilities(ctx context.Context, res statevar.StateVari
 // it returns <defaultProbability> else if there is not yet consensus value it returns <minProbabilityOfTrading>.
 // If there is consensus value and the price is worse than the worse price, we extrapolate using the last 2 price points
 // If the price is within the range, the corresponding probability implied by the offset is returned, scaled, with lower bound of <minProbabilityOfTrading>.
-func getProbabilityOfTrading(bestBid, bestAsk, minPrice, maxPrice num.Decimal, pot *probabilityOfTrading, price num.Decimal, isBid bool, minProbabilityOfTrading num.Decimal) num.Decimal {
+func getProbabilityOfTrading(bestBid, bestAsk, minPrice, maxPrice num.Decimal, pot *probabilityOfTrading, price num.Decimal, isBid bool, minProbabilityOfTrading num.Decimal, offsetOne num.Decimal) num.Decimal {
 	if (isBid && price.GreaterThanOrEqual(bestBid)) || (!isBid && price.LessThanOrEqual(bestAsk)) {
 		return defaultInRangeProbabilityOfTrading
 	}
@@ -173,15 +173,15 @@ func getProbabilityOfTrading(bestBid, bestAsk, minPrice, maxPrice num.Decimal, p
 		if price.LessThan(minPrice) {
 			return minProbabilityOfTrading
 		}
-		return getBidProbabilityOfTrading(bestBid, pot.bidOffset, pot.bidProbability, price, minProbabilityOfTrading)
+		return getBidProbabilityOfTrading(bestBid, pot.bidOffset, pot.bidProbability, price, minProbabilityOfTrading, offsetOne)
 	}
 	if price.GreaterThan(maxPrice) {
 		return minProbabilityOfTrading
 	}
-	return getAskProbabilityOfTrading(bestAsk, pot.askOffset, pot.askProbability, price, minProbabilityOfTrading)
+	return getAskProbabilityOfTrading(bestAsk, pot.askOffset, pot.askProbability, price, minProbabilityOfTrading, offsetOne)
 }
 
-func getAskProbabilityOfTrading(bestAsk num.Decimal, offsets, probabilities []num.Decimal, price, minProbabilityOfTrading num.Decimal) num.Decimal {
+func getAskProbabilityOfTrading(bestAsk num.Decimal, offsets []uint32, probabilities []num.Decimal, price, minProbabilityOfTrading num.Decimal, offsetOne num.Decimal) num.Decimal {
 	// no consensus yet
 	if len(offsets) == 0 {
 		maxDistance := maxDistanceWhenNoConsensusFactor.Mul(bestAsk)
@@ -190,22 +190,22 @@ func getAskProbabilityOfTrading(bestAsk num.Decimal, offsets, probabilities []nu
 		}
 		return minProbabilityOfTrading
 	}
-	offset := price.Sub(bestAsk).Div(bestAsk)
+	offset := uint32(offsetOne.Mul(price.Sub(bestAsk).Div(bestAsk)).Floor().IntPart())
 
 	// if outside the range - extrapolate
-	if offset.GreaterThan(offsets[len(offsets)-1]) {
+	if offset > offsets[len(offsets)-1] {
 		return lexterp(offsets, probabilities, offset, minProbabilityOfTrading)
 	}
 
+	idx := sort.Search(len(offsets), func(i int) bool {
+		return offset < offsets[i]
+	})
 	// linear interpolation
-	interpolatedProbability := linterp(offsets, probabilities, offset, func(i int) bool {
-		r := offset.LessThan(offsets[i])
-		return r
-	}, minProbabilityOfTrading)
+	interpolatedProbability := linterp(offsets, probabilities, offset, idx, minProbabilityOfTrading)
 	return interpolatedProbability
 }
 
-func getBidProbabilityOfTrading(bestBid num.Decimal, offsets, probabilities []num.Decimal, price, minProbabilityOfTrading num.Decimal) num.Decimal {
+func getBidProbabilityOfTrading(bestBid num.Decimal, offsets []uint32, probabilities []num.Decimal, price, minProbabilityOfTrading num.Decimal, offsetOne num.Decimal) num.Decimal {
 	// no consensus yet
 	if len(offsets) == 0 {
 		maxDistance := maxDistanceWhenNoConsensusFactor.Mul(bestBid)
@@ -215,22 +215,22 @@ func getBidProbabilityOfTrading(bestBid num.Decimal, offsets, probabilities []nu
 		return minProbabilityOfTrading
 	}
 
-	offset := bestBid.Sub(price).Div(bestBid)
+	offset := uint32(offsetOne.Mul(bestBid.Sub(price).Div(bestBid)).Floor().IntPart())
 
 	// if outside the range - extrapolate
-	if offset.GreaterThan(offsets[len(offsets)-1]) {
+	if offset > offsets[len(offsets)-1] {
 		return lexterp(offsets, probabilities, offset, minProbabilityOfTrading)
 	}
 
+	idx := sort.Search(len(offsets), func(i int) bool {
+		return offset < offsets[i]
+	})
 	// linear interpolation
-	interpolatedProbability := linterp(offsets, probabilities, offset, func(i int) bool {
-		r := offset.LessThan(offsets[i])
-		return r
-	}, minProbabilityOfTrading)
+	interpolatedProbability := linterp(offsets, probabilities, offset, idx, minProbabilityOfTrading)
 	return interpolatedProbability
 }
 
-func lexterp(offsets, probabilities []num.Decimal, offset, minProbabilityOfTrading num.Decimal) num.Decimal {
+func lexterp(offsets []uint32, probabilities []num.Decimal, offset uint32, minProbabilityOfTrading num.Decimal) num.Decimal {
 	if len(offsets) == 1 {
 		return minProbabilityOfTrading
 	}
@@ -238,8 +238,8 @@ func lexterp(offsets, probabilities []num.Decimal, offset, minProbabilityOfTradi
 	last2 := offsets[len(offsets)-2]
 	probLast := probabilities[len(offsets)-1]
 	probLast2 := probabilities[len(offsets)-2]
-	slopeNum := offset.Sub(last2)
-	slopeDenom := last.Sub(last2)
+	slopeNum := num.DecimalFromInt64(int64(offset - last2))
+	slopeDenom := num.DecimalFromInt64(int64(last - last2))
 	slope := slopeNum.Div(slopeDenom)
 	prob := probLast2.Add(probLast.Sub(probLast2).Mul(slope))
 	scaled := rescaleProbability(prob)
@@ -247,15 +247,13 @@ func lexterp(offsets, probabilities []num.Decimal, offset, minProbabilityOfTradi
 	return prob
 }
 
-func linterp(offsets, probabilities []num.Decimal, priceOffset num.Decimal, searchFunc func(i int) bool, minProbabilityOfTrading num.Decimal) num.Decimal {
-	i := sort.Search(len(offsets), searchFunc)
-
+func linterp(offsets []uint32, probabilities []num.Decimal, priceOffset uint32, i int, minProbabilityOfTrading num.Decimal) num.Decimal {
 	if i >= len(probabilities) {
 		return num.MaxD(minProbabilityOfTrading, rescaleProbability(probabilities[len(probabilities)-1]))
 	}
 	prev := offsets[i-1]
-	size := offsets[i].Sub(prev)
-	ratio := priceOffset.Sub(prev).Div(size)
+	size := offsets[i] - (prev)
+	ratio := num.DecimalFromInt64(int64(priceOffset - prev)).Div(num.DecimalFromInt64(int64(size)))
 	cRatio := num.DecimalFromInt64(1).Sub(ratio)
 	prob := ratio.Mul(probabilities[i]).Add(cRatio.Mul(probabilities[i-1]))
 	scaled := rescaleProbability(prob)